Renamed common.h -> hdzero.h
