+++ src/hdzero.h
