+++ src/sfr_ext.c
@@ -1,4 +1,3 @@
-#include "common.h"
 #include "sfr_ext.h"
 
 /////////////////////////////////////////////////////////////////
